requirements.txt
autopep8==2.3.2
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
et_xmlfile==2.0.0
flake8==7.2.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
